pyproject.toml
[project]
name = "chivalry2-unofficial-server-browser-backend"
version = "1.0.0"
description = "A server browser backend for unofficial chivalry2 servers. Keeps track of live servers with a heartbeat, and allows clients to list them. Probably usable for other unofficial gameservers."
authors = [{name = "Jacob Barber", email = "Jacoby6000@gmail.com"}]
license = {text = "GPL3"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Flask>=2.3.2",
    "Flask-Limiter>=3.3.1",
    "argparse>=1.4.0",
    "gunicorn==23.0.0",
    "pygelf>=0.4.2",
    "PyJWT[crypto]>=2.8.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.3.0",
    "autoflake>=2.2.0",
    "mypy>=1.4.1",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/server_browser_backend"]

[tool.pytest.ini_options]
pythonpath = [ "src" ]
